rearranged modal window components
