--- src/components/backdrop/style.css
@@ -1,5 +1,0 @@
-.Backdrop {
-  position: fixed;
-  inset: 0;
-  background-color: var(--overlay);
-}
