+++ src/components/modal-window/style.css
@@ -1,3 +1,9 @@
+.ModalWindow-backdrop {
+  position: fixed;
+  inset: 0;
+  background-color: var(--overlay);
+}
+
 .ModalWindow {
   position: absolute;
   display: flex;
@@ -7,7 +13,7 @@
   background-color: var(--background);
   width: 75%;
   height: 80%;
-  overflow-y: scroll;
+  overflow-y: auto;
   padding: 24px;
   gap: 24px;
   border-radius: 8px;
